// packages/fe-merchant-pc/src/components/PaperExamine/editTmp/interface.ts
export interface TableDataType {
    analysis: string
    childList: {
        analysis: string
        childList: any[]
        code: string
        customContent: Record<string, any>
        level: number
        merchantCode: string
        operatorCode: string
        optionList: {
            answer: string
            code: string
            isAnswer: false
            questionCode: string
            sort: string
            title: string
        }[]
        parentCode: string
        referenceCount: number
        referenceStatus: number
        storageTime: number
        title: string
        type: number
    }[]
    code: string
    customContent: Record<string, any>
    level: number
    merchantCode: string
    operatorCode: string
    optionList: {
        answer: string
        code: string
        isAnswer: false
        questionCode: string
        sort: string
        title: string
    }[]
    parentCode: string
    referenceCount: number
    referenceStatus: number
    storageTime: number
    title: string
    type: number
}

/**
 * 试题 hook
 */
export interface PaginationType {
    total?: number
    pageSize?: number
    current?: number
}

export interface UseTableSearchType {
    tableData: TableDataType[]
    selectedKeys: Key[]
    pagination: PaginationType
    formCallBack: (values?: FormParamsType) => void
    paginationChangeHandler: PaginationProps['onChange']
    rowSelection: TableRowSelection<TableDataType>
    selectedRowKeys: TableDataType[]
}

/**
 * 试题form类型
 */
export interface FormParamsType {
    titleLike: string
    questionType: number
    questionLevel: number
    discrimination: number
}